Pulled logic from board.vue to composable folder
Board.vue was getting really messy, seperated it into two files
